average_score type changed to float, schema info added to challenge_stats class

+++ src/migrations/20191002151255_alter_table_challenge_stats.sql
@@ -1,5 +1,5 @@
 ALTER TABLE challenge_stats 
-ADD average_score INTEGER;
+ADD average_score FLOAT;
 
 ALTER TABLE challenge_stats 
 ADD solved_count INTEGER;
